all comments are taken into account
