+++ src/components/Paint.vue
@@ -1,12 +1,5 @@
 <template>
-  <div style="color: aquamarine">Paint.vue</div>
-  <!--el-dialog
-    Название="Посмотреть изображение"
-    :visible.sync="dialogJPG"
-    append-to-body
-  >
-    <canvas id="mycanvas" width="940" height="570"></canvas>
-  </el-dialog-->
+<div></div>
 </template>
 
 <script lang="ts">
